Rename codsoft_taskno1 to codsoft_taskno1
As part of my internship, I worked on a supervised machine learning project to classify Iris flower species using Python. The project involved analyzing and predicting the species of an Iris flower (Setosa, Versicolor, or Virginica) based on four key features:
1.Sepal Length
2.Sepal Width
3.Petal Length
4.Petal Width
The goal was to build a classification model that could accurately identify the species of a flower using historical data from the famous Iris dataset.